Updated openstack/openstack Project: openstack-infra/puppet-nodepool  7dd94ddae8f9f4e0244900c5f8f8ff399c279620
Fix duplicated package for build-essential

When installing nodepool on single_node_ci context, it
gives an error for build-essential duplicated declaration
(nodepool and zuul). Zuul is protecting that with an if !defined
declaration, but nodepool is not doing that.
Enclose all the package list in ensure_packages, to protect
from duplication errors.

Change-Id: Ibfc7af81ff8429d592d7d9fe70b06e1f5be4e77f