Updated openstack/openstack Project: openstack/tempest  275f178ab9c9674fbd12d8eca7f141c4b2b7eb1f
Cleanup wording and section for remote access config doc

This commit updates the wording and section separation for the remote
access section of the tempest configuration guide. Previously, the
details about using the validation group felt tacked on and out of
place. This attempts to correct that and make it clear what the
validation group is used for and how that's combined with the network
allocation.

Change-Id: I2f65d25e8f66944347f669dac14d77057a10975a