Updated openstack/openstack Project: openstack/nova-specs  b14091d716d0049c5e33b051f440518e9674cf11
Integrate Castellan for Key Management

Adds specification for integrating Castellan for key management.

Castellan is a key manager interface library that is intended to be
usable with multiple back ends, including Barbican. The Castellan code
is based on the basic key manager interface that resides in Nova and
Cinder. Now that the key manager interface lives in a separate library,
the key manager code can be removed from Nova and Cinder, and Castellan
can be used as the key manager interface instead.

Related: blueprint use-castellan-key-manager
Change-Id: I025995584e74c0b2df00841697714c10d1afb674