Updated openstack/openstack Project: openstack/horizon  5e7cfd5b52a096bf63eafe90013b8ef013632266
Use Magic-Search for Security Groups step in Launch Instance

Security Groups was the last table in the stanaard Launch Instance steps
that didn't use Magic Search for filtering.  This patch establishes the
use of Magic Search and also removes the search bar from the table to
promote accessibility.

The filters are only for Name and Description right now, but this allows
additional filters to be added as necessary.

Change-Id: I5e6098752d1b1d5c9736c103796b533d656092a6
Closes-Bug: 1434710