Updated openstack/openstack Project: openstack/horizon  a9cc43b2b206c7d281955346044d1354a82ee3cd
getIdsOfType now returns all ids if type undefined

This convenience method is used to extract IDs from an
ActionResult object. Often, the caller is only interested in
modified items of a certain type, but sometimes it is useful
to simply get a list of all IDs in the ActionResult.

Partially-Implements: blueprint angular-registry
Change-Id: I7847c24cbc7a6c4039ce0e195083cc619491e62d