Updated openstack/openstack Project: openstack/cinder-specs  bd4f3087dc38ea28f85c1f6c5016fded89ab7397
Integrate Castellan for Key Management

Adds spec for integrating Castellan for key management.

Castellan is a key manager interface library that is intended to be
usable with multiple back ends, including Barbican. The Castellan code
is based on the basic key manager interface that resides in Nova and
Cinder. Now that the key manager interface lives in a separate library,
the key manager code can be removed from Nova and Cinder, and Castellan
can be used as the key manager interface instead.

Related: blueprint use-castellan-key-manager
Change-Id: Ia0e5fd68ede2d506fcb919cd3ec50c4812de4327