Updated openstack/openstack Project: openstack/requirements  1f71cf81c2e3fbe8ceca87f6d65cdc27a689d167
Add krest python client library

The Kaminario REST (krest) is a python client library that provides ORM
like interface for working with Kaminario K2 REST API.

Krest is aimed to provide rapid enablement of managing and monitoring
Kaminario K2 all-flash arrays.

It is needed for OpenStack cinder volume drivers for Kaminario K2.
https://blueprints.launchpad.net/cinder/+spec/kaminario-iscsi-cinder-driver
https://blueprints.launchpad.net/cinder/+spec/kaminario-fc-cinder-driver

Is the library actively maintained?
  Yes(https://github.com/Kaminario/krest.git).
Is the library good code?
  Yes.
Is the library python 3 compatible?
  Yes.
Is the library license compatible?
  Yes.
Is the library already packaged in the distros we target?
  Use pip to manage installation(sudo pip install krest).
Is the function of this library already covered by other
libraries in global-requirements.txt?
  No.

Change-Id: I032d01ca7f0b293be2ce6d2e7916708bd951668a