Updated openstack/openstack Project: openstack-infra/gear  59d29104cb9c370e49b44f313e568bd43b172e1b
Do not encode the type (b) in the job name

Don't encode string type in the job name in handleRequest(). Without
the explicit .decode() before passing to the string formatter it
would end up looking like b"b'<jobname'\t\n...." which made for
working with gear on the other side near impossible because a
decode of the job data would not result in a consumable string
matching expected values.

Now there is an explicit decode prior to formatting the string
data.

Change-Id: Ib2996b84bce719a2f91c166aaa4278c18f89f88f