Updated openstack/openstack Project: openstack-infra/project-config  413db6f185bdb6ad8d837fad17dc15d0c4075556
Create the ARA project

ARA: Ansible Run Analysis makes Ansible Playbook runs easier
to visualize, understand and troubleshoot.

ARA is three things:
- An Ansible callback plugin to record playbook runs into a local or
  remote database
- A CLI client to query the database
- A web interface to visualize the database

The project is an initiative of the RDO community to help parse
the results of CI jobs that are driven by Ansible.

More details are available on the openstack-infra mailing list [1]

Please note that there is no launchpad created for this project.
We would prefer skip an eventual migration and integrate with
StoryBoard directly.

[1]: http://lists.openstack.org/pipermail/openstack-infra/2016-May/004257.html

Change-Id: I39bfae2c4d83c250cb3c8dae1b2c68597f523f45