Updated openstack/openstack Project: openstack/neutron-lbaas  5f7a772732afed0d63132b54e2a9e84f96f07327
Avoid unnecessary debug string evaluation

Logging module will evaluate the arguments and create the string
to be printed only if the right logging level is set. Composing
the string to be printed before passing it to the logging module
wastes cpu cycles, if the set logging level anyways drops that
printout.

Change-Id: I75c29beebb2d59b537090c3e4d92bb7d0978f8f5