Updated openstack/openstack Project: openstack-infra/project-config  a0055f5631d0587a89238f8cae5c3e9b7182f50a
horizon job to use tempest plugin in gate

The horizon tempest plugin is working and horizon integration
has been removed from tempest. It's time to change horizon to use
the horizon tempest plugin in the gate and make the job voting.

This change also removes horizon's use of the
gate-tempest-dsvm-full job as this no longer tests horizon.

Change-Id: I3f2a996f26c462edd968fb961c05cebc83b175a0