Updated openstack/openstack Project: openstack-infra/project-config  7629582c9e27fcf3de16c13ca6a62df35546cd6d
Ironic: fetch tempest plugin from git for non-master

The current setup will always use the tempest plugin from the version of
ironic under test. However, we want to always use the master version of
the plugin, so if the change is not proposed against master, fetch
master from git for the tempest plugin.

Change-Id: I9aeb1ac42a85c31115c25cb7fba883da24b45203