Updated openstack/openstack Project: openstack/nova  2caa45edcf2cf4f9c8ff4439e41dc4acb21ba63a
Fix resource tracking for instances with no numa topology

This fixes a problem in host NUMA node resource tracking when
there is an instance with no numa topology on the same node as
instances with numa topology.

It's triggered while running the resource audit, which ultimately
calls hardware.get_host_numa_usage_from_instance() and assigns
the result to self.compute_node.numa_topology.

The problem occurs if you have a number of instances with numa
topology, and then an instance with no numa topology. When running
numa_usage_from_instances() for the instance with no numa topology
we cache the values of "memory_usage" and "cpu_usage". However,
because instance.cells is empty we don't enter the loop. Since the
two lines in this commit are indented too far they don't get called,
and we end up appending a host cell with "cpu_usage" and
"memory_usage" of zero.   This results in a host numa_topology cell
with incorrect "cpu_usage" and "memory_usage" values, though I think
the overall host cpu/memory usage is still correct.

The fix is to reduce the indentation of the two lines in question
so that they get called even when the instance has no numa topology.
This writes the original host cell usage information back to it.

Change-Id: I7e327b79b731393ed787c4e131dc6d9654f424d0
Closes-Bug: #1590607