Updated openstack/openstack Project: openstack/nova  fc4d9c40370436afe9b114042968ed553ca94814
Adjust field types and defaults on Inventory

An Inventory object has some legitimate constraints on its values
that should be reflected in the object. All numeric values that
have a default are non-negative.

The version of the object doesn't need to be updated because
we've not been passing these things around over RPC.

Change-Id: Iadfa041a419e8a7d6f4e3bb62989a08708aa8290
Partialy-Implements: blueprint generic-resource-pools