Updated openstack/openstack Project: openstack/nova  5cc4f2bd254277264f84721ae9b7ba5944a781e9
Remove unused _vlan_is_disabled test flag

The FakeNetworkAPI was setting the _vlan_is_disabled class member to
False and there was no override of that member anywhere else in the
code. It was checked in one place to determine whether or not to raise a
NotImplementedError. But since it was never True that was never raised.

Change-Id: I830bf1e5c62ec842701fed48b6ecc9ddf7deb2ba