Updated openstack/openstack Project: openstack/nova  b2bbb3276af713ee030b9e7bdb6d49e56adf7b36
deprecate s3 image service config options

The "S3ImageService" class needed config options to set the
expected behavior. This class is not referenced in the code anymore,
as we deprecated S3 in Kilo and removed (most of) the code in Mitaka
with commit 4140eb4 with a proper release note.

This change deprecates the s3 image module related config options.

Change-Id: I65e2973b6660e36c791e1c644cfca9da7519d716