Updated openstack/openstack Project: openstack-infra/project-config  8e8278f0aa690e7a4a8871a66b1126a5293fe519
Clean up proposal slave in a builder

So we can avoid using the jenkins-specific postbuildscript plugin,
move the proposal slave cleanup to a builder.  Run it first so that
it always runs.  Since we're running it first, we can't use it to
clean up our current workspace, so instead, ask it to clean up all
workspaces.

Change-Id: I54a187d643fb76f10cc11411d546c41348b5592e