Updated openstack/openstack Project: openstack/python-neutronclient  521ff7cee562df5105c6404fadb261092af7e342
Add no-shared option to qos-policy-update command

QoS policy can now be updated to be not shared with
other tenants if it was shared before.
To set QoS policy as not shared option "--no-shared"
should be passed to qos-policy-update command.
Option '--shared' works still like it was before so
this change is backward compatible.

Change-Id: I18d32813b3c59cbadfe1f4a1b9ba06725a1d7bb8
Closes-Bug: #1590942