Updated openstack/openstack Project: openstack/neutron  07f9e50f37bfc734066281f417217ade6ebc0bad
OVS: don't throw KeyError when duplicate VLAN tags exist

In _restore_local_vlan_map() we can have two ports with the
same VLAN tag, but trying to remove the second will throw
a KeyError, causing the agent to not start.  Use discard()
instead so we only remove an entry if it's there.

Closes-bug: #1526974
Change-Id: I479c693f490c704c5b6c1462e9ab236684e9c259