Updated openstack/openstack Project: openstack-infra/project-config  726e0344ba4512f9ac6684024dfaacee1ec8d788
Remove copy_puppet_logs publisher

Remove copy_puppet_logs publisher and run a script in
openstack/puppet-openstack-integration to collect logs.
This script is provided by [1] and copied from here.

It also change the puppet-openstack-integration-checkout builder to
clone openstack/puppet-openstack-integration not in ./ but in
./puppet-openstack-integration, so we can use this builder in beaker
jobs, that run for openstck/puppet-* modules.

[1] I625491cdc03e6b3dadff46eafe0811544805e861

Change-Id: I9a5563ded5b99bc216b9b2c7861299f5256456fa