Updated openstack/openstack Project: openstack/python-keystoneclient  b5a39da7ce7712aff8ca6e1da44b3a0617593d0e
Add users functional tests

Adds initial set of functional tests for users. For now, all the tests
are created under a single class. Once we have a gate that runs against
LDAP, we will create a class that only contains readonly tests and a
tox call for it (e.g tox -e functional-readonly).

Change-Id: Ie41872a1ad86db6219dd0af47dbc3e2db46bd878