Updated openstack/openstack Project: openstack-infra/project-config  858227cfc494c3648043c8e690e69a32c1f4491a
Simplify gate tests for 'intel-nfv-ci-tests'

This project is a Tempest plugin used in a 3rd party CI. Only a simple
README is required and there are no unit tests, meaning both the 'docs'
and 'py27' gate tests can be removed. In addition, there's no need to
publish a tarball, as only the source is required. This leaves the PEP8
tests, which still has value, and the tests provided by the project
itself, which a 3rd party CI will run.

Change-Id: Ia00f32baee90594a4c91d872bbce83b9fdc927b7