Updated openstack/openstack Project: openstack/neutron  3281eeb2bb92d746eaea47d4985ddc27af5dead0
Return fixed_ips from port update

fixed_ips has a join relationship with port. During a DB transaction,
it will be acquired once. If the IPAllocation is updated after
acquiring fixed_ips, the port DB model object will still have the
stale value of fixed_ips, which is empty in the case of reported bug.

Expire the fixed_ips of the port can make the next query of port have
the latest fixed_ips, and thus, has the latest fixed_ips in the result
of update port.

Change-Id: I8ca52caf661daa2975cf53212d008eb953d83cc0
Closes-Bug: #1585738