Updated openstack/openstack Project: openstack-infra/storyboard-webclient  5ee64a993b1483f890f80fc6e75cef137b1a4c46
Allow superusers to edit other users successfully

Currently, trying to edit a user other than yourself gives an error
because the email field is unset. This commit fixes that by not trying
to set the email field to NULL, and also correctly updates the cache
so changes are shown correctly without refreshing the page.

Change-Id: I335f24b64c0206dcabadf047fb877a6c7030e890