Updated openstack/openstack Project: openstack/python-novaclient  a70de7e5900fd0e71e49b31e81e3f8c23f0c1c83
Add support for microversion 2.28

microversion 2.28 now returns a dict instead of a string for cpu_info item
returned by GET /os-hypervisors/<id>

Fortunately, the utils.flatten_dict() method checks whether it's a JSON field
so the behaviour keeps to be the same for the CLI output of nova hypervisor-show.

That said, the novaclient API call to HypervisorManager.get() will of course
be returning either a dict or a string for the keyed cpu_info given the
microversion.

Co-Authored-By: Andrey Kurilin <andr.kurilin@gmail.com>
Change-Id: I3a8bfcb7672005430ca99bab0d20e25f48a7e0f6
Implements: blueprint nova-api-hypervsor-cpu-info