Updated openstack/openstack Project: openstack-infra/storyboard-webclient  8533c63cac3bde386a1bfe7ac982531660152287
Only update changed fields when editing tasks

This fixes a bug where two people can open a story detail view, make
changes to different fields in parallel, and only the change made
last survives (as it overwrites the earlier change).

With this patch such a situation would only occur if both people were
changing the same field, which is the expected and correct behaviour.

Change-Id: I09ebaafaef36c067944fa1253f91ea97ba04c8a8