Updated openstack/openstack Project: openstack/oslo.messaging  5690156d9b73e39547dbc3d27b12b41321bb6500
Add query paramereters to TransportURL

This patch proposes to allow TransportURLs to have query parameters
with driver-specific options which may override corresponding
values from a static configuration. As an example, the patch
implements this possibility in the pika driver. If the idea is
approved, it will be possible to also implement it in the other
drivers.

Change-Id: Ibb7e13b4509fde035a3c334cde9bc4c498f92140