Updated openstack/openstack Project: openstack/keystone  53b186306ce7379d869e93ad324326625fd95818
Change LocalUser sql model to eager loading

This patch changes the LocalUser sql model to eager loading. Subquery
loading is eager loading that creates separate SQL statements when the
object is loading, as opposed to a single SQL statement. This is
what was intended and consistent with the other relationships within
the User class.

Change-Id: I48965676ad6a796115caef5e90974cb617243223