Updated openstack/openstack Project: openstack-infra/zuul  490f4aa7d81ce497b7b6ed8854f63ad886f242cb
Increase the Gerrit trailing delay

We're starting to see inconsistent data from gerrit on uploads of
(somewhat) long patch series to nova.  Increase the trailing delay
to give Gerrit time to make itself consistent.  Hopefully we can
reduce this in the future after we work out the GC issues on the
nova repo.

Change-Id: I644fcdc196b737909c2115071027dc2826c21c0b