Updated openstack/openstack Project: openstack-dev/devstack  e13c933d7001faa254b17b2f7d396cfe23a43c3e
Enable admin_domain_scope by default in Tempest

Tempest introduced a new ability to use domain scoped tokens for
identity v3 admin APIs. Since domain scoped tokens can be used
with the base keystone policy used in the gate, and the
pre-provisioned admin user is assigned a role on the domain, turn
the option alway on.

Change-Id: Ib1bb958eee076364b407fc03e77e6882d92147d2
Depends-on: I91ca907992428a5a14fb8d48a4fad105d2906e27