Updated openstack/openstack Project: openstack/taskflow  774d59813417a945fdbe50316786effd1203e8f7
Rename engine analyzer to be named selector

This moves out the engine next to run (or revert)
selection process to a single class that just does
this and moves out the common functions the analyzer
class provided to the runtime object (which all
components can access).

This makes it easier to adjust the selection algorithm
in different ways.

Change-Id: I091c69297a7bff60729791d3ca6c3fae14d6eea5