Updated openstack/openstack Project: openstack/keystonemiddleware  19f9a309cd0babdf1a14a0bc42fb38bf4243df35
Create a Config object

The _conf_get ugliness in auth_token middleware has been around for a
long time now to handle the abstraction from different oslo.config
options and the paste overrides. This logic is now also being needed in
other middlewares. Extract this into a common config object that has a
better interface and is easier to work with.

Change-Id: I8b8a1427bc527e43bb1baec25a881d93df3f93cc