Updated openstack/openstack Project: openstack/keystonemiddleware  1c3eafbb1fa919ffacf9a4b8d0dbc7417982196e
Make audit middleware use common config object

Audit middleware is starting to need the same sort of config handling
that auth_token middleware uses. Make it use the common config object.

The test that is removed is because the config project returns None not
unknown. This is then used correctly, however the test as written is not
useful.

Change-Id: Ic99f6435e01661c2b13dd53d7f9f29357b4e244e